fix: moved data loading logic from +page.server.ts to +page.svelte
