+++ sites/geohub/src/routes/(app)/+page.svelte
@@ -54,8 +54,8 @@
 	<meta name="twitter:description" content={SiteInfo.site_description} />
 	<meta property="og:title" content={title} />
 	<meta property="og:image" content="/api/og?content={content}" />
-	<meta property="og:image:width" content="1288" />
-	<meta property="og:image:height" content="800" />
+	<meta property="og:image:width" content="1200" />
+	<meta property="og:image:height" content="630" />
 	<meta name="twitter:card" content="summary_large_image" />
 	<meta name="twitter:title" content={title} />
 	<meta name="twitter:image" content="/api/og?content={content}" />
