--- sites/geohub/src/routes/(maps)/dashboards/electricity/+page.server.ts
@@ -1,47 +1,0 @@
-import type { PageServerLoad } from './$types';
-import { ELECTRICITY_DATASETS } from './constansts';
-import type { Dataset } from './interfaces';
-
-export const load: PageServerLoad = async ({ fetch }) => {
-	const datasets = ELECTRICITY_DATASETS;
-
-	const hrea: Promise<Dataset>[] = [];
-
-	for (const ds of datasets.hrea) {
-		hrea.push(
-			new Promise<Dataset>((resolve) => {
-				fetch(`/api/datasets/${ds.id}`)
-					.then((res) => res.json())
-					.then((data) => {
-						ds.url = data.properties.url;
-						resolve(ds);
-					});
-			})
-		);
-	}
-
-	const ml: Promise<Dataset>[] = [];
-
-	for (const ds of datasets.ml) {
-		ml.push(
-			new Promise<Dataset>((resolve) => {
-				fetch(`/api/datasets/${ds.id}`)
-					.then((res) => res.json())
-					.then((data) => {
-						ds.url = data.properties.url;
-						resolve(ds);
-					});
-			})
-		);
-	}
-
-	const hreaData = Promise.all(hrea);
-	const mlData = Promise.all(ml);
-
-	return {
-		promises: {
-			hrea: hreaData,
-			ml: mlData
-		}
-	};
-};
